requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.45.0
django-import-export==2.6.1
django-widget-tweaks==1.4.9
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
tablib==3.1.0
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
